refactor: updated path imports after folder structure changes
