+++ src/pages/menu/index.tsx
@@ -1,5 +1,5 @@
-import { UserLayout } from '@/components';
 import { MenuView } from '@/features';
+import { UserLayout } from '@/shared/layout';
 
 import type { NextPageWithLayout } from '../_app';
