+++ src/pages/sign-in/index.tsx
@@ -1,5 +1,5 @@
-import { UserLayout } from '@/components';
 import { SignInView } from '@/features';
+import { UserLayout } from '@/shared/layout';
 
 import type { NextPageWithLayout } from '../_app';
